pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-debate-arena"
version = "0.1.0"
description = "CLI-first multi-agent AI debate arena with JSON export and static web viewer."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "httpx>=0.27,<1.0",
  "jinja2>=3.1,<4.0",
  "pyyaml>=6.0,<7.0",
  "rich>=13.0,<14.0",
  "typer>=0.12,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "pytest-asyncio>=0.23,<1.0",
  "ruff>=0.11,<1.0",
]
web = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.34,<1.0",
]

[project.scripts]
debate = "debate.cli:app"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["debate*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
  "live: real API integration tests",
  "slow: long-running end-to-end tests",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
